PRS.jl Documentation

PRS.jl Documentation

PlinkReader

LDMatrix

ldpred

lassosum

utils

lassosum

Implementation of ldpred, EB, and lassosum in Julia.

PlinkReader

PRS.PlinkReader — Method

PlinkReader(path; markerIndex = false, sampleIndex = false)

Structure to access PLINK .bed, .bim, and .fam files at path.

Arguments

markerIndex::Bool should index marker -> idx be created?

sampleIndex::Bool should index sample -> idx be created?

PRS.Sample — Type

Sample(fid, iid, father, mother, sex)

PLINK sample info

PRS.nsamples — Method

nsamples(p::PlinkReader)

Number of samples in Plink file.

PRS.samples — Method

samples(p::PlinkReader)

Return Array of Sample structs for samples in Plink file.

PRS.sample_index — Method

sample_index(p::PlinkReader, iid::String)

Return index of sample iid in PlinkReader. Uses index if available, otherwise linear search.

PRS.Marker — Type

Marker(chrom, id, cm, pos, a1, a2)

PLINK marker info

PRS.nmarkers — Method

nmarkers(p::PlinkReader)

Number of markers in Plink file.

PRS.markers — Method

markers(p::PlinkReader)

Return array of Marker structs for markers in Plink file.

PRS.marker_index — Method

marker_index(p::PlinkReader, id::String)

Return index of marker id in PlinkReader. Uses index if available, otherwise linear search.

PRS.markersDF — Method

markersDF(p::PlinkReader)

Get data frame with marker info.

Returns data frame with columns

Chrom

Name

cM

Pos

A1

A2

Idx

Base.getindex — Method

getindex(p::PlinkReader, s::Int, m::Int)

Retrieve genotype of sample s and marker m.

Plink convention for representation of genotypes is as follows:

0b00 Hom1

0b01 Het

0b10 missing

0b11 Hom2

PRS.genotypeMatrix — Function

genotypeMatrix(p::PlinkReader, markerIdx, sampleIdx = nothing;
               normalize = true)

Create genotype matrix (samples x markers) for markerIdx and sampleIdx from PlinkReader. Here genotype is expressed as count of alternative allele (not as in Plink convention).

Replace missing genotypes with mean genotype. If normalize = true, normalize markers to mean μ = 0 and standard deviation σ = 1.

LDMatrix

PRS.LDMatrix — Method

LDMatrix(p::PlinkReader, mIdx0, mIdx1; alpha = 0.9, window = posWindow(1_000_000))

Compute LDMatrix from genotypes in Plink file.

PRS.LDMatrix — Method

LDMatrix(path::AbstractString)

Load LDMatrix from .bim and .lds files at path.

PRS.markers — Method

markers(ld::LDMatrix)

Get array of Marker structs for all markers in LDMatrix ld.

PRS.markersDF — Method

markersDF(ld::LDMatrix)

Get data frame with marker info.

Returns data frame with columns

Chrom

Name

cM

Pos

A1

A2

Idx

PRS.save — Method

save(ld::LDMatrix, path::AbstractString)

Save LDMatrix to path.

PRS.ldscore — Method

ldscore(ld::LDMatrix)

Compute LD scores for markers in LDMatrix ld.

ldpred

PRS.ldpred_gibbs — Method

ldpred_gibbs(z, D, p, σ2, μ0; n_burnin = 100, n_iter = 500, verbose = false)

Run LDpred Gibbs sampler.

Arguments

z::Vector: z-scores

D::LDMatrix: LD matrix

p::Real : proportion of variants deemed to be causal

σ2::Real: $Nh²/M$ as estimated from LDscore regression (estimate_h2, or mean(z^2)/mean(lds)) Note: prior variance of non-null component of $μ$, $σ2_μ = σ2/p$

μ0::Vector: starting estimate (e.g. from infinite model)

PRS.estimate_h2 — Method

estimate_h2(z, lds, n)

Estimate total heritability $h^2$ for markers using LD score regression with intercept forced to 1.

Arguments

lds::Vector: LD score as computed by ldscore(LDMatrix)

n::Integer: effective number of samples ($4 n_0*n_1/(n_0+n_1)$ for case-control)

PRS.estimate_neff — Method

estimate_neff(beta, sebeta, freq)

Estimate effective number of samples in case-control study.

Effective number of samples is total number of samples for a cohort with same number of samples and controls, that would result in the observed sebeta.

freq can be minor or major allele freq, as only freq*(1-freq) is being used.

lassosum

PRS.z2cor — Method

z2cor(z, n)

Convert GWAS z-value to phenotype-genotype correlation coefficient.

Missing docstring.

Missing docstring for elnetg!(X, r, λ, s, β = zeros(Float64, length(r)); maxiter = 10_000, thresh = 1e-4). Check Documenter's build log for details.

PRS.elnetg_path — Method

elnetg_path(X, r, λs, s;
            maxiter = 10_000, thresh = 1e-4)

Solve elastic net for path along λs using warm starts.

Arguments

X::Matrix nsubj x nmarkers normalized genotype matrix column normalized (μ = 0, σ = 1)

r::Vector nmarkers x 1 vector of correlation coefficients between phenotype and genotypes

λs::Vector shrinkage parameter path for 1-norm in decreasing order

s::Real shrinkage parameter for 2-norm (LD)

β::Vector warm start (in) and result (out) for solution vector

maxiter::Int maximum number of iterations

thresh::Real maximum change in β to be called converged

utils

PRS.snp_join — Method

snp_join(df1, df2; on = [:Chrom, :Pos],
         alleles1 = [:A1_1, :A2_1], alleles2 = [:A1_2, :A2_2],
         matchcol = :sign)

Join data frames df1 and df2 and match variants.

Data frames df1 and df2 are first matched by columns specified in on. Then column matchcol is set to

matchcol == 1 if values for alleles1 and alleles2 match

matchcol == -1 if values for alleles1 and alleles2 are swapped

matchcol == 0 otherwise